requirements.txt
# 核心依赖
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
click>=8.0.0
pyyaml>=6.0.0

# AI 集成
google-generativeai>=0.3.0

# 数据存储
notion-client>=2.0.0

# 机器人平台
python-telegram-bot>=20.0.0
itchat-uos==1.4.1
aiohttp>=3.8.0

# 工具库
asyncio-mqtt>=0.11.0  # 可选：MQTT支持
requests>=2.28.0
urllib3>=1.26.0
schedule>=1.2.0  # 定时任务调度器
python-crontab>=3.0.0  # 高级定时任务支持

# 调试工具（可选）
# 注意：使用Ngrok调试功能需要单独安装ngrok命令行工具
# 下载地址：https://ngrok.com/download 